Fix shellhost infiniteloop issue (ssh.exe -t user@ip dir) (#292)
Fix for PowerShell/Win32-OpenSSH#892
Fixed warning messages in latest code.

+++ contrib/win32/win32compat/socketio.c
@@ -195,7 +195,7 @@ socketio_WSARecv(struct w32_io* pio, BOOL* completed, int len)
 		wsabuf.buf = pio->read_details.buf;
 
 	if (len)
-		wsabuf.len = min(len, wsabuf.len);
+		wsabuf.len = min((ULONG)len, wsabuf.len);
 
 	ret = WSARecv(pio->sock, &wsabuf, 1, NULL, &recv_flags, &pio->read_overlapped, &WSARecvCompletionRoutine);
 	if (ret == 0) {
